got the tables working and got some USER data confirmed with testing, adding Profile for user and working on table relations

+++ db.js
@@ -10,12 +10,12 @@ const sequelize = new Sequelize(
     }
 )
 
-async function synceDb(sequelize, force = false){
+async function synceDb(sequelize, options){
+    const {force, alter } = options
     try {
-        if (force)
-            await sequelize.sync({force: true})
-        else
-            await sequelize.sync()
+        if (force) await sequelize.sync({force: true})
+        else if (alter) await sequelize.sync({ alter: true})
+        else await sequelize.sync()
     } catch (err){
         console.log(err)
     }
